refactor(chartData): move funtions to a new utility file to avoid code repetition
