+++ src/utils/stackedCardData.js
@@ -1,22 +1,5 @@
 import getDate from './getDate.js';
-
-const globalColors = [
-  ['rgba(255, 0, 0, 0.1)', 'rgba(255, 0, 0, 0.5)'],
-  ['rgba(0, 255, 0, 0.1)', 'rgba(0, 255, 0, 0.5)'],
-  ['rgba(0, 0, 255, 0.1)', 'rgba(0, 0, 255, 0.5)'],
-  ['rgba(244, 244, 65, 0.1)', 'rgba(244, 244, 65, 0.5)'],
-  ['rgba(244, 124, 65, 0.1)', 'rgba(244, 124, 65, 0.5)'],
-  ['rgba(65, 244, 211, 0.1)', 'rgba(65, 244, 211, 0.5)'],
-  ['rgba(160, 65, 244, 0.1)', 'rgba(160, 65, 244, 0.5)'],
-];
-let globalColorIndex = 0;
-
-function getColor() {
-  const color = globalColors[globalColorIndex];
-  globalColorIndex = (globalColorIndex + 1) % globalColors.length;
-
-  return color;
-}
+import { getLabels, buildChartDataSets } from './chartUtils.js';
 
 function getActivityData(activity, dateTypeSelector, dayOfWeek) {
   if (activity.type === 'updateCard') {
@@ -35,48 +18,6 @@ function getActivityData(activity, dateTypeSelector, dayOfWeek) {
   }
 
   return null;
-}
-
-function getLabels(activities) {
-  return activities.map((element) => element.date)
-    .filter((value, index, self) => self.indexOf(value) === index)
-    .sort();
-}
-
-function countCardsByWeek(activities, week) {
-  return activities.filter((activity) => activity.date === week).length;
-}
-
-function buildChartDataSet(filteredActivities, labels) {
-  if (filteredActivities.length === 0) {
-    return {
-      label: [],
-      data: [],
-    };
-  }
-  const datasetName = filteredActivities[0].list.name;
-  const datasetValues = labels.map((label) => countCardsByWeek(filteredActivities, label))
-    .reduce((a, b, index) => [...a, b + (a[index - 1] || 0)], []);
-
-  const colors = getColor();
-
-  return {
-    label: datasetName,
-    data: datasetValues,
-    fill: true,
-    backgroundColor: colors[0],
-    borderColor: colors[1],
-    cubicInterpolationMode: 'monotone',
-  };
-}
-
-function buildChartDataSets(activities, labels, listIds) {
-  return listIds.map(
-    (listId) => buildChartDataSet(
-      activities.filter((activity) => listId === activity.list.id),
-      labels
-    )
-  );
 }
 
 function fillRetroactively(activities, listIds) {
